requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-telegram-bot>=20.0
ta>=0.10.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
joblib>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

